[fset-1384] Set initial AC passmarks - initial refactor
