+++ app/controllers/AssessmentCentrePassMarkSettingsController.scala
@@ -16,9 +16,8 @@
 
 package controllers
 
-import model.Commands.Implicits.assessmentCentrePassMarkSettingsResponseFormat
-import model.PassmarkPersistedObjects.AssessmentCentrePassMarkSettings
-import model.PassmarkPersistedObjects.Implicits._
+import model.persisted.assessmentcentre.AssessmentCentrePassMarkSettings
+import model.Commands.Implicits._
 import play.api.libs.json.Json
 import play.api.mvc.Action
 import repositories._
