+++ app/model/Commands.scala
@@ -22,17 +22,14 @@ import model.CandidateScoresCommands.CandidateScoresAndFeedback
 import model.CandidateScoresCommands.Implicits._
 import model.Exceptions.{ NoResultsReturned, TooManyEntries }
 import model.OnlineTestCommands.Implicits._
-import model.OnlineTestCommands.TestResult
-import model.PassmarkPersistedObjects.{ AssessmentCentrePassMarkInfo, AssessmentCentrePassMarkScheme }
-import model.PassmarkPersistedObjects.Implicits._
+import model.persisted.assessmentcentre.{ AssessmentCentrePassMarkInfo, AssessmentCentrePassMarkScheme }
 import org.joda.time.{ DateTime, LocalDate, LocalTime }
 import play.api.libs.json._
 
 import scala.language.implicitConversions
-import model.command.{ AssessmentCentre, ProgressResponse }
-import model.exchange.passmarksettings.Phase1PassMarkSettings
+import model.command.{ ProgressResponse }
 import model.persisted.{ QuestionnaireAnswer, QuestionnaireQuestion }
-import model.report.{ CandidateProgressReportItem, QuestionnaireReportItem }
+import model.report.{ QuestionnaireReportItem }
 import reactivemongo.bson.BSONDocumentReader
 
 //scalastyle:off
